services/users_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
django-prometheus>=2.3
celery>=5.4
Pillow>=10.0
